tests/requirements.txt
pyyaml
termcolor